+++ components/Component.js
@@ -35,11 +35,11 @@ function removeListener(removeEventListenerHandle) {
     removeEventListenerHandle();
 }
 
-function checkCompatibleComponent(componentsContext, el) {
+function checkCompatibleComponent(globalComponentsContext, el) {
     var component = el._w;
     while(component) {
         var id = component.id;
-        var newComponentDef = componentsContext.$__componentsById[id];
+        var newComponentDef = globalComponentsContext.$__componentsById[id];
         if (newComponentDef && component.$__type == newComponentDef.$__component.$__type) {
             break;
         }
@@ -168,9 +168,9 @@ function onBeforeNodeDiscarded(node) {
     return eventDelegation.$__handleNodeDetach(node);
 }
 
-function onBeforeElUpdated(fromEl, key, componentsContext) {
-    if (componentsContext && key) {
-        var preserved = componentsContext.$__preserved[key];
+function onBeforeElUpdated(fromEl, key, globalComponentsContext) {
+    if (key) {
+        var preserved = globalComponentsContext.$__preserved[key];
 
         if (preserved === true) {
             // Don't morph elements that are associated with components that are being
@@ -181,14 +181,14 @@ function onBeforeElUpdated(fromEl, key, componentsContext) {
             // We may need to destroy a Component associated with the current element
             // if a new UI component was rendered to the same element and the types
             // do not match
-            checkCompatibleComponent(componentsContext, fromEl);
+            checkCompatibleComponent(globalComponentsContext, fromEl);
         }
     }
 }
 
-function onBeforeElChildrenUpdated(el, key, componentsContext) {
-    if (componentsContext && key) {
-        var preserved = componentsContext.$__preservedBodies[key];
+function onBeforeElChildrenUpdated(el, key, globalComponentsContext) {
+    if (key) {
+        var preserved = globalComponentsContext.$__preservedBodies[key];
         if (preserved === true) {
             // Don't morph the children since they are preserved
             return MORPHDOM_SKIP;
@@ -196,8 +196,8 @@ function onBeforeElChildrenUpdated(el, key, componentsContext) {
     }
 }
 
-function onNodeAdded(node, componentsContext) {
-    eventDelegation.$__handleNodeAttach(node, componentsContext.$__out);
+function onNodeAdded(node, globalComponentsContext) {
+    eventDelegation.$__handleNodeAttach(node, globalComponentsContext.$__out);
 }
 
 var componentProto;
@@ -549,19 +549,18 @@ Component.prototype = componentProto = {
                 while(targetEl) {
                     var id = targetEl.id;
 
-                    if (id) {
-                        fromEl = fromEls[id];
-                        if (fromEl) {
-                            morphdom(
-                                fromEl,
-                                targetEl,
-                                componentsContext,
-                                onNodeAdded,
-                                onBeforeElUpdated,
-                                onBeforeNodeDiscarded,
-                                onNodeDiscarded,
-                                onBeforeElChildrenUpdated);
-                        }
+                if (id) {
+                    fromEl = fromEls[id];
+                    if (fromEl) {
+                        morphdom(
+                            fromEl,
+                            targetEl,
+                            globalComponentsContext,
+                            onNodeAdded,
+                            onBeforeElUpdated,
+                            onBeforeNodeDiscarded,
+                            onNodeDiscarded,
+                            onBeforeElChildrenUpdated);
                     }
 
                     targetEl = targetEl.nextSibling;