+++ components/legacy/renderer-legacy.js
@@ -1,5 +1,6 @@
-var componentLookup = require('../util').$__componentLookup;
-var ComponentsContext = require('../ComponentsContext');
+var getComponentsContext = require('../ComponentsContext').$__getComponentsContext;
+var componentsUtil = require('../util');
+var componentLookup = componentsUtil.$__componentLookup;
 var registry = require('../registry');
 var modernRenderer = require('../renderer');
 var resolveComponentKey = modernRenderer.$__resolveComponentKey;
@@ -73,6 +74,7 @@ function createRendererFunc(templateRenderFunc, componentProps) {
             }
         }
 
+        var componentsContext = getComponentsContext(out);
         id = id || componentsContext.$__nextComponentId();
 
         if (registry.$__isServer && typeName) {
@@ -157,7 +159,7 @@ function createRendererFunc(templateRenderFunc, componentProps) {
         if (!component) {
             fakeComponent = {};
         } else {
-            componentState = component.$__rawState;
+            componentState = component.$__rawState || componentState;
         }
 
         var templateInput = getTemplateData ?